requirements.txt
paho-mqtt
finnhub-python